requirements.txt
sagemaker
gradio
python-dotenv